Revamped Pub/Subs
All publications are handled by a single publisher for the time being,
due to conflicts with the way I was keeping track of all the instances
of GameEvent I had stored in the GameEventRegistry and the way
Monobehaviours... behave with the Awake() function. Trying to make
GameEvent Monobehaviours made a lot of problems for me.

Also I'm pretty proud of my new dynamic LenkedObjectRow object. Funny
how problems I just couldn't figure out last month are almost trivial
for me to solve now.